changed rate calculation
Instead of relying on random() to fire the events, informant
now is guaranteed to send one event per X requests, where X is
the statsd_sample_rate set in the config. The side effect is that
informant will only be able to track up to sys.maxint requests per
second (this should be sufficient for most use cases).